+++ thundermod/CHANGELOG.md
@@ -1,3 +1,16 @@
+# 1.1.1
+
+- Failed to save changelog in 1.1.0
+
+# 1.1.0
+
+- Add item and furniture (combined) blacklist to prevent errors (suits break ads permanently in vanilla (v72))
+- Add customizable ad slogans, defaults to vanilla.
+
+# 1.0.1
+
+- Fix dll version stuff
+
 # 1.0.0
 
 - Adding a server side config. LethalConfig is helpful but not required.